refactor IzstarInfo to IzstarWiki
