--- src/IzstarInfo/data.json
@@ -1,49 +1,0 @@
-{
-  "紫微": {
-    "name": "紫微星",
-    "yinYang": "阴",
-    "fiveElement": "土",
-    "galaxy": "中天",
-    "transformation": "尊贵",
-    "elementColor": "黄色",
-    "energyColor": "紫光",
-    "occupation": "官禄主",
-    "position": "众星枢纽，长五行，孕万物",
-    "alias": "帝王星，老板星，俸禄星",
-    "characteristics": [
-      "看到紫微星，相信很多读者都知道他是一颗帝王星，但并不是说紫微星的人能当帝王，而是说紫微星人的个性具有“王者”的特质，比如说控制欲、占有欲较强，比较喜欢一些华而不实的物品和称赞，典型的要面子不要里子的类型。也正因为如此，紫微星会给人比较强的距离感。紫微星有着很多领导型人格所需要的特质：目标长远，大局观强、好为人师等等。",
-      "这是一颗非常“吃资源”的星曜，如果给予足够的支持和条件去培养这颗星，让紫微星具有和自身气场所匹配的能力，紫微星就会发挥出他强大的控场和规划能力。但是如果自身的认知水平不足，没有足够能知识和能力去支撑这颗星，他就会变得比较眼高手低，高不成低不就了。",
-      "所以假如你的命宫主星是紫微星的话，一定要扩展自己知识的广度，把自身基础打牢固。这样在工作和生活中就非常容易成为一个受人尊敬的人。另外，紫微星要特别注意被人“捧杀”，因为他特别容易迷失在别人的吹捧之中。"
-    ],
-    "celestialConfigurations": [
-      {
-        "primaryStar": "紫微",
-        "secondaryStars": "天府",
-        "characteristics": [
-          "紫微星和天府星都是帝星，双帝星的组合会加重帝王那种控制欲和占有欲，喜欢一切情况尽在掌握的感觉。但是内心会出现要虚荣还是要务实的冲突，所以在遇到抉择的时候反而难以果断的下决定。",
-          "要注意审视自己的能力是否能匹配上自己的眼光。紫府对奢侈品有比较重的情节，在吃穿用度上要求都颇高。自尊心和虚荣心都极强的紫府如果生于清贫之家成长路上就比较辛苦，因为紫府很喜欢攀比，就算因为天府的关系不喜欢表现出来，但是内心是会默默的去比较自己和周围人的状况。",
-          "求学阶段就比较喜欢竞选班干部之类的管理职务，喜欢安排别人。如善加引导，是比较容易成为优秀的管理者。相对于紫微星单星，紫府同宫格的人的欲望更加强烈，因为紫府同宫的格局下，福德宫一定是 贪狼星，所以相对的，爱好也会比较广泛。",
-          "由于考虑得比较长远，想得也比较多的关系，紫府的人也比较容易在哲学方面有比较高的造诣。这个也和紫府与人的距离感有关，在紫府成长初期通常会觉得自己应该是人群中的C位。久而久之，大家就不爱带着他玩儿了，这也是紫府孤独感的来源。这种与人相处的距离感会让部分紫府反思，在反思中顿悟。一个顿悟的紫府会给人一种肃然起敬的感觉。"
-        ]
-      },
-      {
-        "primaryStar": "紫微",
-        "secondaryStars": "七杀",
-        "characteristics": [
-          "紫微星配上杀破狼中的任何一颗星，都会让紫微星变得比较接地气，让紫微星变得比较务实。七杀星 会让有点摇摆不定的紫微星变得比较坚定，不容易受周围环境或人影响自己的决定。",
-          "一个兼具执行力的皇帝一旦方向走对了，那是很可怕的。所以，紫微七杀适合在某个专业领域深耕下去，假以时日，一般都会在某个领域有所建树。但是他一旦摆烂起来，也非常难办，因为紫微七杀本不太愿意听别人的意见，又比较孤傲，所以要改变他的想法，是件难度很大的事情。",
-          "从上面的特点不难看出，紫微七杀一旦被PUA，要纠正也是非常困难的事情。因此紫微七杀要多多增加自己的见闻，在自己不熟悉的领域中要谨慎，以免被人怂恿着把自己陷入进退两难的情况。"
-        ]
-      },
-      {
-        "primaryStar": "紫微",
-        "secondaryStars": "破军",
-        "characteristics": [
-          "紫微星和破军星组合，就像皇帝得到了军队，所以做起事情来有一种摧枯拉朽的感觉。所以从小比较叛逆的、不太服管。但是紫微破军因为迁移宫一定是 天相星 的关系，也就有了点“窝里横”的意味。就是说不熟的时候感觉他还挺会顾及别人感受的，熟络以后反而不太给面子。",
-          "这个组合表现出来的气度很大方、不拘小节、不计较。这种豪迈大方的个性本身很容易让大家跟着他玩，然后紫微破军那种死要面子活受罪的个性非常容易让他打肿脸充胖子，所以社交圈子对于紫微破军来说影响力还蛮大的。",
-          "如果周围都是酒肉朋友，奸诈之人，那紫微破军可以说会过得比较难受了。所以如果你的命宫主星是紫微破军，要学会拒绝，不要为了面子硬着头皮撑。紫微破军在运动和艺术类的天赋都比较高，但是正经学习的话，一般都比较不上心。假如你小孩是紫微破军，你有想让他坐得住静下心来学习，是需要费一番功夫的。"
-        ]
-      }
-    ]
-  }
-}
