--- src/IzstarInfo/index.ts
@@ -1,1 +1,0 @@
-export * from "./IzstarInfo";
