+++ src/IzstarWiki/IzstarWiki.type.ts
