+++ src/IzstarWiki/doc/紫微.md
@@ -1,0 +1,86 @@
+## 紫微星 <Badge type="warning" text="命宫出现概率≈1.28%" />
+
+<table class="star-card">
+  <tr>
+    <td>阴阳</td>
+    <td>阴 🌑</td>
+    <td>五行</td>
+    <td>己土 ⛰️</td>
+    <td>斗分</td>
+    <td>中天星系</td>
+  </tr>
+  <tr>
+    <td>化气</td>
+    <td>尊贵</td>
+    <td>五行色</td>
+    <td>黄色</td>
+    <td>能量色</td>
+    <td>紫光</td>
+  </tr>
+  <tr>
+    <td>职业</td>
+    <td>官禄主</td>
+    <td>职务</td>
+    <td>众星枢纽，长五行，孕万物</td>
+    <td>别号</td>
+    <td>帝王星，老板星，俸禄星</td>
+  </tr>
+</table>
+
+### 紫微星特性
+
+看到紫微星，相信很多读者都知道他是一颗 `帝王星`，但并不是说紫微星的人能当帝王，而是说紫微星人的个性具有“王者”的特质，比如说控制欲、占有欲较强，比较喜欢一些华而不实的物品和称赞，典型的要面子不要里子的类型。也正因为如此，紫微星会给人比较强的距离感。紫微星有着很多领导型人格所需要的特质：目标长远，大局观强、好为人师等等。
+
+这是一颗非常“吃资源”的星曜，如果给予足够的支持和条件去培养这颗星，让紫微星具有和自身气场所匹配的能力，紫微星就会发挥出他强大的控场和规划能力。但是如果自身的认知水平不足，没有足够能知识和能力去支撑这颗星，他就会变得比较眼高手低，高不成低不就了。
+
+所以假如你的命宫主星是 `紫微星` 的话，一定要扩展自己知识的广度，把自身基础打牢固。这样在工作和生活中就非常容易成为一个受人尊敬的人。另外，紫微星要特别注意被人“捧杀”，因为他特别容易迷失在别人的吹捧之中。
+
+### 紫微星组合
+
+---
+
+#### 紫微天府 <Badge type="warning" text="命宫出现概率≈1.47%" />
+
+紫微星和 `天府星` 都是帝星，双帝星的组合会加重帝王那种控制欲和占有欲，喜欢一切情况尽在掌握的感觉。但是内心会出现要虚荣还是要务实的冲突，所以在遇到抉择的时候反而难以果断的下决定。
+
+要注意审视自己的能力是否能匹配上自己的眼光。紫府对奢侈品有比较重的情节，在吃穿用度上要求都颇高。自尊心和虚荣心都极强的紫府如果生于清贫之家成长路上就比较辛苦，因为紫府很喜欢攀比，就算因为天府的关系不喜欢表现出来，但是内心是会默默的去比较自己和周围人的状况。
+
+求学阶段就比较喜欢竞选班干部之类的管理职务，喜欢安排别人。如善加引导，是比较容易成为优秀的管理者。相对于紫微星单星，紫府同宫格的人的欲望更加强烈，因为紫府同宫的格局下，福德宫一定是 `贪狼星`，所以相对的，爱好也会比较广泛。
+
+由于考虑得比较长远，想得也比较多的关系，紫府的人也比较容易在哲学方面有比较高的造诣。这个也和紫府与人的距离感有关，在紫府成长初期通常会觉得自己应该是人群中的 C 位。久而久之，大家就不爱带着他玩儿了，这也是紫府孤独感的来源。这种与人相处的距离感会让部分紫府反思，在反思中顿悟。一个顿悟的紫府会给人一种肃然起敬的感觉。
+
+#### 紫微七杀 <Badge type="warning" text="命宫出现概率≈1.36%" />
+
+紫微星配上杀破狼中的任何一颗星，都会让紫微星变得比较接地气，让紫微星变得比较务实。`七杀星` 会让有点摇摆不定的紫微星变得比较坚定，不容易受周围环境或人影响自己的决定。
+
+一个兼具执行力的皇帝一旦方向走对了，那是很可怕的。所以，紫微七杀适合在某个专业领域深耕下去，假以时日，一般都会在某个领域有所建树。但是他一旦摆烂起来，也非常难办，因为紫微七杀本不太愿意听别人的意见，又比较孤傲，所以要改变他的想法，是件难度很大的事情。
+
+从上面的特点不难看出，紫微七杀一旦被 PUA，要纠正也是非常困难的事情。因此紫微七杀要多多增加自己的见闻，在自己不熟悉的领域中要谨慎，以免被人怂恿着把自己陷入进退两难的情况。
+
+#### 紫微破军 <Badge type="warning" text="命宫出现概率≈1.39%" />
+
+紫微星和 `破军星` 组合，就像皇帝得到了军队，所以做起事情来有一种摧枯拉朽的感觉。所以从小比较叛逆的、不太服管。但是紫微破军因为迁移宫一定是 `天相星` 的关系，也就有了点“窝里横”的意味。就是说不熟的时候感觉他还挺会顾及别人感受的，熟络以后反而不太给面子。
+
+这个组合表现出来的气度很大方、不拘小节、不计较。这种豪迈大方的个性本身很容易让大家跟着他玩，然后紫微破军那种死要面子活受罪的个性非常容易让他打肿脸充胖子，所以社交圈子对于紫微破军来说影响力还蛮大的。
+
+如果周围都是酒肉朋友，奸诈之人，那紫微破军可以说会过得比较难受了。所以如果你的命宫主星是紫微破军，要学会拒绝，不要为了面子硬着头皮撑。紫微破军在运动和艺术类的天赋都比较高，但是正经学习的话，一般都比较不上心。假如你小孩是紫微破军，你有想让他坐得住静下心来学习，是需要费一番功夫的。
+
+#### 紫微贪狼 <Badge type="warning" text="命宫出现概率≈2.83%" />
+
+这个组合被称为 `桃花犯主`，因为 `贪狼星` 化气桃花，是紫微斗数的第一桃花星。这个组合在 太微赋 中有一句 `桃花犯主至淫` 的说法，让不少人在看到这句话的时候看紫微贪狼的眼神都变了。其实这个真的是误会大了。
+
+紫微贪狼自带社交属性，喜欢各种新奇人事物，好奇心重，所以只要遇到他感兴趣的领域，学什么都快，但是比较难专精。通常紫微贪狼都会给人一种看起来比较风流的感觉，不管男女，都会比较有才。而在古代，有才的人都会流连勾栏，所以才落了个“至淫”的名头。
+
+贪狼星代表着欲望，所以他会加重紫微星的占有欲和控制欲，这让紫微贪狼眼里容不得沙子。但，他还有那种只许州官放火，不许百姓点灯的霸道。在处理男女关系的问题上，紫微贪狼尤其容易造成不可挽回的后果。
+
+在“吹牛”这件事情上，紫贪也是及其有天赋的，而且他还禁不起激，所以常常自己给自己挖坑跳。有些事情明知不可为而为之，那叫有魄力。但有些事明明知道自己搞不定，却在一众人的吹捧下觉得自己行了，那就是傻了。
+
+#### 紫微天相 <Badge type="warning" text="命宫出现概率≈1.44%" />
+
+古籍有云 `紫微天相，为臣不忠，为子不孝`。可以说描述得非常之夸张了，紫微天相的人看到这句话不由得瑟瑟发抖：“可千万别让我妈/领导看到这句话”。其实大可不必。因为社会形态的原因，当时喜欢顺从，听话而且没有自己想法的人，所以一切有自己想法，顶嘴，不服从管教的人都会被冠以不忠不孝之名。
+
+`天相星` 是一颗非常平衡的星曜，他什么都能来一点儿，不管是策划，管理，执行还是组织，规划，社交，统统在他的能力范围。在他和紫微星搭配在一起以后，会让他“敢于”去将自己的想法表达或者实践出来，并且因为有紫微星的“撑腰”，让他敢于对不合理说不。
+
+这种特质让紫微天相的“前途”直接和伯乐挂了勾，因为紫微天相需要一个没有束缚的展现自我的平台，如果遇到一个妒忌心很强的领导或者控制欲很强的父母，那紫微天相可就难受了。本身这个组合就是出现在 `天罗地网` 位，所以受到环境和心境的影响会很大。
+
+紫微减低了天相星那种左右摇摆的个性，而天相减低了紫微的距离感。这个组合就像皇帝(紫微星)有了印(天相星)，将军有了虎符，可以说是非常平衡的一个组合了。
